fix(rmm-eac2): fix MISRA C:2012 Rule 10.3 in EXTRACT usage
Fixes
Event misra_c_2012_rule_10_3_violation:
- Implicit conversion of "(read_id_aa64mmfr0_el1() & 0xfUL) >> 0U"
  from essential type "unsigned 64-bit int" to different or
  narrower essential type "signed 64-bit int".
- Implicit conversion of "(read_id_aa64dfr0_el1() & 0xf00UL) >> 8UL"
  from essential type "unsigned 64-bit int" to different or
  narrower essential type "unsigned 32-bit int".
- Implicit conversion of "(esr & 0x1f0000UL) >> 16" from essential
  type "unsigned 64-bit int" to different or narrower essential
  type "unsigned 32-bit int".
- Implicit conversion of "(esr & 0xc00000UL) >> 22" from essential
  type "unsigned 64-bit int" to different or narrower essential
  type "unsigned 32-bit int".
- Implicit conversion of "(vtr & 0x1fUL) >> 0UL" from essential
  type "unsigned 64-bit int" to different or narrower essential
  type "unsigned 32-bit int".

+++ lib/gic/src/gic.c
@@ -71,7 +71,8 @@ void gic_get_virt_features(void)
 	unsigned long nr_pri_bits;
 
 	/* Number of implemented List registers, minus 1 */
-	gic_virt_feature.nr_lrs = EXTRACT(ICH_VTR_EL2_LIST_REGS, vtr);
+	gic_virt_feature.nr_lrs = (unsigned int)EXTRACT(ICH_VTR_EL2_LIST_REGS,
+							vtr);
 	assert(gic_virt_feature.nr_lrs < ICH_MAX_LRS);
 
 	/* Number of virtual preemption bits implemented */
